change double to bigdecimal

+++ src/main/java/pl/devdmin/core/order/ShippingCalculationStrategy.java
@@ -1,5 +1,7 @@
 package pl.devdmin.core.order;
 
+import java.math.BigDecimal;
+
 public interface ShippingCalculationStrategy {
-    double getShippingCost();
+    BigDecimal getShippingCost();
 }
